Move GpgKeypair into CpEnv namespace
