+++ lib/cp_env.rb
@@ -5,6 +5,7 @@ require "open-uri"
 require "open3"
 require 'base64'
 require 'json'
+require 'tmpdir'
 
 class CpEnv
 end
@@ -14,7 +15,5 @@ require File.join(File.dirname(__FILE__), "cp_env", "pipeline")
 require File.join(File.dirname(__FILE__), "cp_env", "terraform")
 require File.join(File.dirname(__FILE__), "cp_env", "namespace_deleter")
 require File.join(File.dirname(__FILE__), "cp_env", "kubeconfig")
+require File.join(File.dirname(__FILE__), "cp_env", "gpg_keypair")
 require File.join(File.dirname(__FILE__), "cp_env", "gitops_gpg_keypair")
-
-# TODO: Move gpg stuff into CpEnv
-require File.join(File.dirname(__FILE__), "gpg_keypair")
