+++ spec/gpg_keypair_spec.rb
@@ -1,10 +1,10 @@
 require "spec_helper"
 
-describe GpgKeypair do
+describe CpEnv::GpgKeypair do
   subject(:keypair) { described_class.new }
 
   it "instantiates" do
-    expect(keypair).to be_a(GpgKeypair)
+    expect(keypair).to be_a(CpEnv::GpgKeypair)
   end
 
   it "generates a keypair" do
